backend/requirements.txt
Django
psycopg2-binary
djangorestframework
cryptography
django-cors-headers
djangorestframework-simplejwt
markdown
django-filter
gunicorn
django-debug-toolbar
boto3
django-rest-knox